+++ vegafusion-datafusion-udfs/src/udfs/datetime/mod.rs
@@ -4,9 +4,9 @@ pub mod datetime_components;
 pub mod datetime_format;
 pub mod epoch_to_timestamptz;
 pub mod str_to_timestamptz;
-pub mod timestamp_to_timestamptz;
+pub mod to_utc_timestamp;
 pub mod timestamptz_to_epoch;
-pub mod timestamptz_to_timestamp;
+pub mod from_utc_timestamp;
 pub mod timeunit;
 
 use crate::udfs::datetime::str_to_timestamptz::datetime_strs_to_timestamp_millis;
