+++ vegafusion-sql/src/dialect/mod.rs
@@ -627,8 +627,8 @@ impl Dialect {
                 "isnan",
                 "isfinite",
                 "pow",
-                "timestamp_to_timestamptz",
-                "timestamptz_to_timestamp",
+                "to_utc_timestamp",
+                "from_utc_timestamp",
                 "date_to_timestamptz",
                 "epoch_ms_to_timestamptz",
                 "str_to_timestamptz",
